spread js code into modules
