+++ src/blocks/popup/_opened/popup_opened.css
@@ -1,6 +1,6 @@
 .popup_opened {
   visibility: visible;
   opacity: 1;
-  transition: visibility, opacity .2s ease-in-out;
+  transition: visibility .2s, opacity .2s ease-in-out;
 }
