+++ src/index.html
@@ -11,7 +11,7 @@
 <body class="page">
   <main class="content">
     <header class="header">
-      <img src="./images/mesto-logo.svg" class="header__logo" alt="Mesto logo.">
+      <img src="<%=require('./images/mesto-logo.svg')%>" class="header__logo" alt="Mesto logo.">
     </header>
     <section class="profile">
       <div class="profile__container">
@@ -37,16 +37,18 @@
   <div class="popup popup__profile">
     <div class="popup__container">
       <h2 class="popup__heading">Редактировать профиль</h2>
-      <form class="form" name="edit-form">
+      <form class="form" name="edit-form" novalidate>
         <fieldset class="form__fieldset">
           <label for="name" class="form__field">
             <input class="form__item form__item_type_name" type="text" id="name" name="name" placeholder="Ваше имя"
-              required minlength="2" maxlength="40">
+              required minlength="2" maxlength="40" pattern="^[а-яА-ЯёЁa-zA-Z\s-]+$"
+              data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
             <span class="form__item-error name-error"></span>
           </label>
           <label for="activity" class="form__field">
             <input class="form__item form__item_type_job" type="text" id="activity" name="activity" placeholder="О себе"
-              required minlength="2" maxlength="200">
+              required minlength="2" maxlength="200" pattern="^[а-яА-ЯёЁa-zA-Z\s-]+$"
+              data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
             <span class="form__item-error activity-error"></span>
           </label>
         </fieldset>
@@ -58,16 +60,18 @@
   <div class="popup popup__add-photo">
     <div class="popup__container">
       <h2 class="popup__heading">Новое место</h2>
-      <form class="form card-form" name="card-form" id="card-form">
+      <form class="form card-form" name="card-form" id="card-form" novalidate>
         <fieldset class="form__fieldset">
           <label for="name" class="form__field">
             <input class="form__item form__item_type_photo" type="text" id="name" name="photo-name"
-              placeholder="Название" required minlength="2" maxlength="30">
+              placeholder="Название" required minlength="2" maxlength="30" pattern="^[а-яА-ЯёЁa-zA-Z\s-]+$"
+              data-error-message="Разрешены только латинские, кириллические буквы, знаки дефиса и пробелы">
             <span class="form__item-error name-error"></span>
           </label>
           <label for="place" class="form__field">
             <input class="form__item form__item_type_link" type="link" id="place" name="link"
-              placeholder="Ссылка на картинку" required>
+              placeholder="Ссылка на картинку"
+              data-error-message="Вставьте ссылку на картинку" required>
             <span class="form__item-error place-error"></span>
           </label>
         </fieldset>
